moved storyListing to product
