+++ .github/CODEOWNERS
@@ -16,3 +16,4 @@ src/products/event/ @department-of-veterans-affairs/ap-team
 src/products/eventListing/ @department-of-veterans-affairs/ap-team
 src/products/staffProfile/ @department-of-veterans-affairs/ap-team
 src/products/newsStory/ @department-of-veterans-affairs/ap-team
+src/products/storyListing/ @department-of-veterans-affairs/ap-team
