+++ src/data/queries/index.ts
@@ -41,7 +41,7 @@ import * as QuestionAnswer from '@/data/queries/questionAnswer'
 import * as ReactWidget from '@/data/queries/reactWidget'
 import * as ResourcesSupport from '@/data/queries/resourcesSupport'
 import * as StaticPathResources from '@/data/queries/staticPathResources'
-import * as StoryListing from '@/data/queries/storyListing'
+import * as StoryListing from '@/products/storyListing/query'
 import * as SupportServices from '@/data/queries/supportServices'
 import * as Table from '@/data/queries/table'
 import * as VamcEhr from '@/data/queries/vamcEhr'
