+++ src/pages/[[...slug]].tsx
@@ -43,7 +43,7 @@ import { PressRelease as FormattedPressRelease } from '@/types/formatted/pressRe
 import { PressReleaseListing as FormattedPressReleaseListing } from '@/types/formatted/pressReleaseListing'
 import { ResourcesSupport as FormattedResourcesSupport } from '@/types/formatted/resourcesSupport'
 import { StaffProfile as FormattedStaffProfile } from '@/products/staffProfile/formatted-type'
-import { StoryListing as FormattedStoryListing } from '@/types/formatted/storyListing'
+import { StoryListing as FormattedStoryListing } from '@/products/storyListing/formatted-type'
 import { VetCenter as FormattedVetCenter } from '@/types/formatted/vetCenter'
 import { HealthCareLocalFacility as FormattedHealthCareLocalFacility } from '@/types/formatted/healthCareLocalFacility'
 import { VamcSystem as FormattedVamcSystem } from '@/types/formatted/vamcSystem'
@@ -62,7 +62,7 @@ import { PressReleaseListing } from '@/templates/layouts/pressReleaseListing'
 import { PreviewCrumb } from '@/templates/common/preview'
 import { ResourcesSupport } from '@/templates/layouts/resourcesSupport'
 import { StaffProfile } from '@/products/staffProfile/template'
-import { StoryListing } from '@/templates/layouts/storyListing'
+import { StoryListing } from '@/products/storyListing/template'
 import { VetCenter } from '@/templates/layouts/vetCenter'
 import { Wrapper } from '@/templates/layouts/wrapper'
 import { HealthCareLocalFacility } from '@/templates/layouts/healthCareLocalFacility'
